refactor: AI Code Review 반영 및 오타 수정
