+++ qrust-api/src/main/java/com/qrust/config/ObjectMapperConfig.java
@@ -1,4 +1,4 @@
-package com.qrust.confing;
+package com.qrust.config;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
