+++ qrust-api/src/main/java/com/qrust/qrcode/infrastructure/zxing/QrCodeGeneratorImpl.java
@@ -5,7 +5,7 @@ import com.google.zxing.BarcodeFormat;
 import com.google.zxing.client.j2se.MatrixToImageWriter;
 import com.google.zxing.common.BitMatrix;
 import com.google.zxing.qrcode.QRCodeWriter;
-import com.qrust.qrcode.appliaction.QrCodeGenerator;
+import com.qrust.qrcode.application.QrCodeGenerator;
 import com.qrust.qrcode.utils.AesUtil;
 import com.qrust.qrcode.domain.entity.vo.QrCodeData;
 import java.awt.image.BufferedImage;
